Move libraries into lib directory
