+++ lib/libfst/fastlz.c
