+++ lib/libfst/fastlz.h
