+++ lib/libfst/lz4.c
