+++ lib/libfst/meson.build
@@ -8,6 +8,7 @@ libfst_dependencies = [
     zlib_dep,
     bzip2_dep,
     judy_dep,
+    thread_dep,
 ]
 
 libfst = static_library(
