+++ lib/libghw/libghw.c
