+++ lib/libjrb/meson.build
@@ -1,0 +1,9 @@
+libjrb = static_library(
+    'jrb',
+    'jrb.c',
+)
+
+libjrb_dep = declare_dependency(
+    link_with: libjrb,
+    include_directories: '.',
+)
