+++ lib/liblxt/lxt2_write.h
@@ -43,8 +43,6 @@ extern "C" {
 #define ftello ftell
 #endif
 
-#include "wavealloca.h"
-
 #define LXT2_WR_HDRID (0x1380)
 #define LXT2_WR_VERSION (0x0001)
