+++ lib/liblxt/lxt_write.c
