+++ lib/liblxt/lxt_write.h
