+++ lib/libvzt/vzt_read.c
