+++ lib/libvzt/vzt_read.h
