+++ lib/libvzt/vzt_write.h
