+++ lib/meson.build
@@ -1,0 +1,5 @@
+subdir('libfst')
+subdir('libghw')
+subdir('libjrb')
+subdir('liblxt')
+subdir('libvzt')
